Update imports and overall package structure
Used the structure of the shablona project [1] as a model for imports. Now the
tests can query the output as well as the functions in the project for unit tests.
However, running the pyinseq.py from the command line would return the error:

```
SystemError: Parent module '' not loaded, cannot perform relative import
```

Therefore, added `run.py` outside of the package as an entry point for the package.
Not sure if this is a long-term solution or a temporary hack.

[1] shablona: https://github.com/uwescience/shablona